requirements.txt
PySide6==6.7.0
Pillow==10.2.0
mysql-connector-python==8.3.0
bcrypt==4.1.2